spark-apps/discount/requirements.txt
pyspark
psycopg2-binary
kafka-python
